Chord Naming Quiz - COMPLETE
Complete rewrite of "BackendDictionary" - now named "ChordDictionary".
I didn't delete the original in case someone was using it for something.

Chord naming quiz is officially complete! This works for 3, 4, and 5
note chords. I also added a button that clears and resets all user
input tiles. There’s a temporary answer label that displays the correct
set of notes for the given chord. This is just for debugging and needs
to be removed before release.